finish cleanup-latex.py

+++ py3/cleanup-latex.py
@@ -6,8 +6,12 @@ files = os.listdir()
 
 p = re.compile(".+\.(aux|log|synctex\.gz|toc)")
 
-for file in matched_files:
-    filesize_sum += os.stat(file).st_size / 1024 # file size in kB
-    os.remove(file)
+matched_files = []
+filesize_sum = 0
+for file in files:
+    if re.match(p, file):
+        matched_files.append(file)
+        filesize_sum += os.stat(file).st_size / 1024 # file size in kB
+        os.remove(file)
 
 print("{0} Dateien gelöscht, {1} kB Speicher freigegeben.".format(len(matched_files), filesize_sum))
